Domain structures lie in a separate module
Extract them into main module ('lib.rs').
