+++ src/serialize.rs
@@ -14,20 +14,16 @@ lazy_static! {
     static ref REQ_ID: Arc<AtomicUsize> = Arc::new(AtomicUsize::new(1));
 }
 
-impl<'a> Serialize for ::method::MethodParams<'a> {
+impl<'a> Serialize for ::MethodParams<'a> {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
         where S: Serializer
     {
         match self.0 {
-            ::method::Method::ClientVersion => {
-                serializer.serialize_some(&method("web3_clientVersion"))
-            }
-            ::method::Method::EthSyncing => serializer.serialize_some(&method("eth_syncing")),
-            ::method::Method::EthBlockNumber => {
-                serializer.serialize_some(&method("eth_blockNumber"))
-            }
-            ::method::Method::EthAccounts => serializer.serialize_some(&method("eth_accounts")),
-            ::method::Method::EthGetBalance => {
+            ::Method::ClientVersion => serializer.serialize_some(&method("web3_clientVersion")),
+            ::Method::EthSyncing => serializer.serialize_some(&method("eth_syncing")),
+            ::Method::EthBlockNumber => serializer.serialize_some(&method("eth_blockNumber")),
+            ::Method::EthAccounts => serializer.serialize_some(&method("eth_accounts")),
+            ::Method::EthGetBalance => {
                 serializer.serialize_some(&method_params("eth_getBalance", self.1))
             }
         }
